Moved netty and mina classes to package
