+++ server/src/main/java/handling/MinaMapleServerHandler.java
@@ -27,8 +27,8 @@ import handling.cashshop.CashShopOperationHandlers;
 import handling.cashshop.CashShopServer;
 import handling.channel.handler.InterServerHandler;
 import handling.channel.handler.PlayerHandler;
-import handling.session.MinaMaplePacketDecoder;
-import handling.session.MinaSession;
+import handling.session.mina.MinaMaplePacketDecoder;
+import handling.session.mina.MinaSession;
 import handling.world.WorldServer;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.mina.common.IdleStatus;
