+++ server/src/main/java/handling/NettyMapleServerHandler.java
@@ -6,9 +6,9 @@ import handling.cashshop.CashShopOperationHandlers;
 import handling.cashshop.CashShopServer;
 import handling.channel.handler.InterServerHandler;
 import handling.channel.handler.PlayerHandler;
-import handling.session.NettyMaplePacketDecoder;
-import handling.session.NettyMaplePacketEncoder;
-import handling.session.NettySession;
+import handling.session.netty.NettyMaplePacketDecoder;
+import handling.session.netty.NettyMaplePacketEncoder;
+import handling.session.netty.NettySession;
 import handling.world.WorldServer;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
