+++ server/src/main/java/handling/session/SocketProvider.java
@@ -1,0 +1,9 @@
+package handling.session;
+
+import handling.PacketProcessor;
+
+public interface SocketProvider {
+    void initSocket(int channel, int port, PacketProcessor.Mode mode);
+
+    void unbindAll();
+}
