+++ server/src/main/java/handling/session/netty/NettyMaplePacketEncoder.java
@@ -1,4 +1,4 @@
-package handling.session;
+package handling.session.netty;
 
 import client.MapleClient;
 import io.netty.buffer.ByteBuf;
